core: extract foundation package; fix pipeline_module wrong-direction import
Pulls the shared idea→premise→spine→world-bible→chapter-plan stages, the
act-slicer, the sanity check, the artifact writer, and the dataclasses
(WorldBible, PlanEntry, Character, WorldState + render_world_state) into a
new core/ package. Each variant pipeline now imports foundation from one
canonical place instead of cross-importing between siblings.

Tangle fix: pipeline_module.py was reaching into pipeline_ws.py for
_build_foundation. Both _build_foundation (now build_foundation) and
build_world_bible live in core.foundation; all three pipelines depend on
core directly.

Variant-specific code untouched:
- story.py keeps run_enhance_chapter + run_generate_story_so_far (variant A).
- world_state.py keeps run_init/draft/advance_world_state (variant B).
- story_module.py keeps WriteStory + DraftChapter + StoryOutline (variant C).
  These three modules become generators/{baseline,world_state,dspy_module}.py
  when the StoryGenerator protocol lands (Phase 6/7).

UI decoupling deferred: foundation run_* functions still call ui.review_answer
directly. The orchestrator will inject a reviewer callback in a follow-up so
foundation can be exercised without an interactive shell.

All 38 tests pass; artifact and pov_check/qa imports unaffected.
