+++ pipeline_module.py
@@ -1,17 +1,19 @@
-"""Story pipeline variant that drafts chapters via the :mod:`story_module` dspy.Module.
+"""Variant C (dspy.Module) pipeline: foundation → WriteStory module.
 
-Everything up to and including the world bible is the existing interactive
-pipeline (reused verbatim via :func:`pipeline_ws._build_foundation`). From
-there a non-interactive two-stage ``dspy.Module`` — outline → draft each
-chapter, with no rolling summary — takes over. See :class:`story_module.WriteStory`.
+Shares the foundation with variants A and B via
+:func:`core.foundation.build_foundation`, then a non-interactive two-stage
+``dspy.Module`` — outline → draft each chapter, with no rolling summary —
+takes over. See :class:`story_module.WriteStory`.
+
+Once the registry refactor (Phase 7) lands, the body of :func:`write` becomes
+the entry point of ``generators/dspy_module.py``.
 """
 
 import logging
 
 from _compat import observe
-from artifact import update_artifact
-from pipeline_ws import _build_foundation  # shared idea→premise→spine→world-bible steps
-from story import sanity_check
+from core.artifact import update_artifact
+from core.foundation import build_foundation, sanity_check
 from story_module import WriteStory
 
 logger = logging.getLogger(__name__)
@@ -42,7 +44,7 @@ def _record_chapters(output_file: str, chapters) -> None:
 @observe()
 def write(idea: str, title: str, output_file: str, number_of_chapters: int = 7) -> None:
     """Run the module-based story pipeline end to end."""
-    updated_idea, story_title, spine, world_bible = _build_foundation(
+    updated_idea, story_title, spine, world_bible = build_foundation(
         idea, title, output_file
     )
